requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
peewee==3.14.10
pip==21.1.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
setuptools==57.0.0
six==1.16.0
telebot==0.0.4
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wheel==0.36.2